copilot_orchestra/pyproject.toml
[project]
name = "copilot-orchestra"
version = "0.1.0"
description = "Multi-agent AI code review platform built on the GitHub Copilot SDK"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "github-copilot-sdk>=0.1.25",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",
    "aiofiles>=24.1.0",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
markers = [
    "integration: marks tests that require a running Copilot CLI (deselect with -m 'not integration')",
]
testpaths = ["tests"]
log_cli = true
log_cli_level = "INFO"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "S", "B"]
ignore = [
    "S101",  # assert in tests
    "S603",  # subprocess without shell — we validate paths separately
    "B008",  # FastAPI Depends() in function signature
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S", "B"]

[tool.coverage.run]
source = ["backend"]
omit = ["tests/*", "backend/main.py"]

[tool.coverage.report]
show_missing = true
fail_under = 70
